gitlab-ci: Use a specific MSVC version on Windows
Teach jobs to load the environment for a specific MSVC version, and tag
them accordingly.  This allows runners to have multiple versions of MSVC
installed without requiring us to compile on the newer ones right away.

The runners provide the VS 2019 location in a `VS160COMNTOOLS`
environment variable.  Use it to avoid requiring a specific VS edition.

Issue: #595