Add tests for divergence, vorticity, and q criterion
The gradient filter has the ability to compute divergence, vorticity,
and q criterion. However, none of the gradient tests were actually
testing the result of these computations. At best, one of the tests was
checking that the arrays were created, but never actually looking at the
values.

For one of the gradient tests, this change computes these 3 values and
checks the results, which look reasonable compared to ParaView.